fix(refactor): flaky shortcut test cases
