+++ tests/shortcuts/bound-actions-layout.spec.ts
@@ -1,0 +1,283 @@
+import { test, expect, Page } from '../../playwright';
+import {
+  createCollection,
+  createRequest,
+  openRequest as openRequestBase,
+  closeAllCollections,
+  createFolder,
+  openCollection,
+  selectRequestPaneTab
+} from '../utils/page';
+
+const modifier = process.platform === 'darwin' ? 'Meta' : 'Control';
+const collectionName = 'kb-collection';
+const baseRequests = ['req-1', 'req-2', 'req-3', 'req-4', 'req-5', 'req-6', 'req-7', 'req-8', 'req-9'];
+
+const setupBoundActionsData = async (page: Page, createTmpDir: (prefix: string) => Promise<string>) => {
+  await closeAllCollections(page);
+  const path = await createTmpDir('kb-collection-path');
+  await createCollection(page, collectionName, path);
+
+  await createFolder(page, 'kb-folder', collectionName, true);
+  await createFolder(page, 'kb-draft-folder', collectionName, true);
+  await createFolder(page, 'kb-terminal-folder', collectionName, true);
+};
+
+const checkIfRequestExists = async (page: Page, requestName: string) => {
+  await openCollection(page, collectionName);
+  const request = page.getByTestId('collections').locator('.collection-item-name').filter({ hasText: requestName }).first();
+  return (await request.count()) > 0;
+};
+
+const openRequest = async (...args: Parameters<typeof openRequestBase>) => {
+  const [page, targetCollectionName, requestName] = args;
+  if (
+    targetCollectionName === collectionName
+    && baseRequests.includes(requestName)
+    && !(await checkIfRequestExists(page, requestName))
+  ) {
+    await createRequest(page, requestName, targetCollectionName);
+  }
+
+  return openRequestBase(...args);
+};
+
+const openKeybindingsTab = async (page: Page) => {
+  await page.getByRole('button', { name: 'Open Preferences' }).click();
+  await page.getByRole('tab', { name: 'Keybindings' }).click();
+  await expect(page.locator('.section-header').filter({ hasText: 'Keybindings' })).toBeVisible();
+};
+
+/**
+ * Close the Preferences tab by clicking its close button.
+ * Using the close button avoids depending on any keyboard shortcut that may
+ * have just been reconfigured.
+ */
+const closePreferencesTab = async (page: Page) => {
+  const prefTab = page.locator('.request-tab').filter({ hasText: 'Preferences' });
+  await prefTab.hover();
+  await prefTab.getByTestId('request-tab-close-icon').click({ force: true });
+  await expect(prefTab).not.toBeVisible({ timeout: 8000 });
+};
+
+const closeTabByName = async (page: any, name: string | RegExp) => {
+  const tab = page.locator('.request-tab').filter({ hasText: name });
+  await tab.click();
+  await tab.hover();
+  await tab.getByTestId('request-tab-close-icon').click({ force: true });
+  await expect(tab).not.toBeVisible({ timeout: 2000 });
+};
+
+const openFolderSettingsTab = async (page: Page, folderName: string) => {
+  await openCollection(page, collectionName);
+  const folderRow = page.locator('.collection-item-name').filter({ hasText: folderName }).first();
+  await expect(folderRow).toBeVisible({ timeout: 5000 });
+  await folderRow.dblclick();
+  await expect(page.locator('.request-tab').filter({ hasText: folderName })).toBeVisible({ timeout: 3000 });
+};
+
+const reopenClosedTab = async (page: Page, shortcut: () => Promise<void>, expectedTabName: string | RegExp) => {
+  for (let attempt = 0; attempt < 3; attempt++) {
+    await page.locator('.request-tab').first().click();
+    await page.waitForTimeout(150);
+    await shortcut();
+    const reopenedTab = page.locator('.request-tab').filter({ hasText: expectedTabName });
+    if ((await reopenedTab.count()) > 0) {
+      await expect(reopenedTab).toBeVisible({ timeout: 3000 });
+      return;
+    }
+    await page.waitForTimeout(200);
+  }
+
+  await expect(page.locator('.request-tab').filter({ hasText: expectedTabName })).toBeVisible({ timeout: 5000 });
+};
+
+const remapKeybinding = async (
+  page: Page,
+  action: string,
+  pressShortcut: () => Promise<void>
+) => {
+  await openKeybindingsTab(page);
+  const row = page.getByTestId(`keybinding-row-${action}`);
+  await expect(row).toBeVisible({ timeout: 5000 });
+  await row.scrollIntoViewIfNeeded();
+  await row.hover();
+  const editButton = row.getByTestId(`keybinding-edit-${action}`);
+  const keybindingInput = page.getByTestId(`keybinding-input-${action}`);
+
+  if (await editButton.isVisible().catch(() => false)) {
+    await editButton.click({ force: true });
+  } else {
+    await row.click({ force: true });
+    if (await editButton.isVisible().catch(() => false)) {
+      await editButton.click({ force: true });
+    }
+  }
+
+  await expect(keybindingInput).toBeVisible({ timeout: 5000 });
+
+  await page.keyboard.press('Backspace');
+  await pressShortcut();
+  await closePreferencesTab(page);
+};
+
+const getTabIndex = async (page: Page, name: string) => {
+  const tabs = page.locator('.request-tab .tab-label');
+  const count = await tabs.count();
+  for (let i = 0; i < count; i++) {
+    const text = (await tabs.nth(i).innerText()).trim();
+    if (text.includes(name)) {
+      return i;
+    }
+  }
+
+  return -1;
+};
+
+// ─── Tests ────
+
+test.describe('Shortcut Keys - BOUND_ACTIONS', () => {
+  test.beforeEach(async ({ page, createTmpDir }) => {
+    await page.locator('[data-app-state="loaded"]').waitFor({ timeout: 5000 });
+    await setupBoundActionsData(page, createTmpDir);
+  });
+
+  test.afterAll(async ({ page }) => {
+    await closeAllCollections(page);
+  });
+
+  test.describe('LAYOUT', () => {
+    test.describe('SHORTCUT: Change Layout', () => {
+      test('default Cmd/Ctrl+J change layout orientation', async ({ page, createTmpDir }) => {
+        await openRequest(page, 'kb-collection', 'req-5', { persist: true });
+
+        // Press Cmd/Ctrl+J to change layout
+        await page.keyboard.press(`${modifier}+KeyJ`);
+
+        await expect(
+          page.getByTestId('response-layout-toggle-btn')
+        ).toHaveAttribute('title', 'Switch to horizontal layout', { timeout: 2000 });
+
+        // Press Cmd/Ctrl+J to change layout
+        await page.keyboard.press(`${modifier}+KeyJ`);
+
+        await expect(
+          page.getByTestId('response-layout-toggle-btn')
+        ).toHaveAttribute('title', 'Switch to vertical layout', { timeout: 2000 });
+
+        // Press Cmd/Ctrl+J to change layout
+        await page.keyboard.press(`${modifier}+KeyJ`);
+
+        await expect(
+          page.getByTestId('response-layout-toggle-btn')
+        ).toHaveAttribute('title', 'Switch to horizontal layout', { timeout: 2000 });
+      });
+
+      test('customized Alt+Shift+Y change layout orientation', async ({ page, createTmpDir }) => {
+        await page.keyboard.down('Alt');
+        await page.keyboard.down('KeyY');
+        await page.keyboard.up('KeyY');
+        await page.keyboard.up('Alt');
+
+        // Remap changeLayout to Alt+D
+        await openKeybindingsTab(page);
+        const row = page.getByTestId('keybinding-row-changeLayout');
+        await row.hover();
+        await page.getByTestId('keybinding-edit-changeLayout').click();
+        await expect(page.getByTestId('keybinding-input-changeLayout')).toBeVisible({ timeout: 2000 });
+
+        await page.keyboard.down('Backspace');
+
+        await page.keyboard.down('Alt');
+        await page.keyboard.down('Shift');
+        await page.keyboard.down('KeyY');
+        await page.keyboard.up('KeyY');
+        await page.keyboard.up('Shift');
+        await page.keyboard.up('Alt');
+
+        await openRequest(page, 'kb-collection', 'req-5', { persist: true });
+
+        await page.keyboard.down('Alt');
+        await page.keyboard.down('Shift');
+        await page.keyboard.down('KeyY');
+        await page.keyboard.up('KeyY');
+        await page.keyboard.up('Shift');
+        await page.keyboard.up('Alt');
+
+        await expect(
+          page.getByTestId('response-layout-toggle-btn')
+        ).toHaveAttribute('title', 'Switch to vertical layout', { timeout: 2000 });
+
+        // Press Cmd/Ctrl+J to change layout
+        await page.keyboard.down('Alt');
+        await page.keyboard.down('Shift');
+        await page.keyboard.down('KeyY');
+        await page.keyboard.up('KeyY');
+        await page.keyboard.up('Shift');
+        await page.keyboard.up('Alt');
+
+        await expect(
+          page.getByTestId('response-layout-toggle-btn')
+        ).toHaveAttribute('title', 'Switch to horizontal layout', { timeout: 2000 });
+
+        // Press Cmd/Ctrl+J to change layout
+        await page.keyboard.down('Alt');
+        await page.keyboard.down('Shift');
+        await page.keyboard.down('KeyY');
+        await page.keyboard.up('KeyY');
+        await page.keyboard.up('Shift');
+        await page.keyboard.up('Alt');
+
+        await expect(
+          page.getByTestId('response-layout-toggle-btn')
+        ).toHaveAttribute('title', 'Switch to vertical layout', { timeout: 2000 });
+      });
+    });
+
+    test.describe('SHORTCUT: Open Preferences', () => {
+      test('default Cmd/Ctrl+, open preferences', async ({ page }) => {
+        await page.keyboard.down('Alt');
+        await page.keyboard.down('KeyY');
+        await page.keyboard.up('KeyY');
+        await page.keyboard.up('Alt');
+
+        // Press Cmd/Ctrl+J to change layout
+        await page.keyboard.down(modifier);
+        await page.keyboard.down('Comma');
+        await page.keyboard.up('Comma');
+        await page.keyboard.up(modifier);
+
+        await expect(page.locator('.request-tab').filter({ hasText: 'Preferences' })).toBeVisible({ timeout: 3000 });
+      });
+
+      test('customized Cmd/Ctrl+P open preferences', async ({ page }) => {
+        // Remap openPreferences to Ctrl+P
+        await openKeybindingsTab(page);
+        const row = page.getByTestId('keybinding-row-openPreferences');
+        await row.hover();
+        await page.getByTestId('keybinding-edit-openPreferences').click();
+        await expect(page.getByTestId('keybinding-input-openPreferences')).toBeVisible({ timeout: 2000 });
+
+        await page.keyboard.down('Backspace');
+
+        await page.keyboard.down(modifier);
+        await page.keyboard.down('KeyP');
+        await page.keyboard.up('KeyP');
+        await page.keyboard.up(modifier);
+
+        await page.keyboard.down('Alt');
+        await page.keyboard.down('KeyY');
+        await page.keyboard.up('KeyY');
+        await page.keyboard.up('Alt');
+
+        // Press Cmd/Ctrl+J to change layout
+        await page.keyboard.down(modifier);
+        await page.keyboard.down('KeyP');
+        await page.keyboard.up('KeyP');
+        await page.keyboard.up(modifier);
+
+        await expect(page.locator('.request-tab').filter({ hasText: 'Preferences' })).toBeVisible({ timeout: 3000 });
+      });
+    });
+  });
+});
